# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Fleet Control API
  version: 0.2.1
  description: |
    Public web API consumed by the Fleet web UI. The API aggregates multiple
    device roles (audio players, video/CEC, Zigbee, cameras) and exposes a
    consistent contract for UI integrations. Schemas align with the live device
    capabilities matrix to ensure client compatibility.
  contact:
    name: Fleet Platform Team
    email: platform@fleet.example
servers:
  - url: /api
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  headers:
    CorrelationId:
      description: Correlation identifier for tracing requests end-to-end.
      schema:
        type: string
    RateLimitLimit:
      description: Maximum number of requests allowed in the current window.
      schema:
        type: integer
        format: int32
    RateLimitRemaining:
      description: Remaining requests in the current window.
      schema:
        type: integer
        format: int32
    RetryAfterSeconds:
      description: Seconds until clients may retry after a 429 response.
      schema:
        type: integer
        format: int32
...
  responses:
    UnauthorizedError:
      description: Authentication failed or credentials missing.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            default:
              value:
                code: UNAUTHENTICATED
                message: Authentication token missing or invalid.
                correlationId: 00000000-0000-0000-0000-000000000000
    ForbiddenError:
      description: Authenticated user does not have permission to access the resource.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundError:
      description: Requested resource does not exist.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: One or more request parameters failed validation.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            invalid:
              value:
                code: INVALID_REQUEST
                message: Provided payload failed validation.
                correlationId: 11111111-1111-1111-1111-111111111111
    RateLimitError:
      description: Request rate limit exceeded.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
        x-ratelimit-limit:
          $ref: '#/components/headers/RateLimitLimit'
        x-ratelimit-remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        retry-after:
          $ref: '#/components/headers/RetryAfterSeconds'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadGatewayError:
      description: Upstream device returned an invalid response or is unreachable.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    GatewayTimeoutError:
      description: Upstream device timed out while processing the request.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Unexpected server error occurred.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotImplementedError:
      description: Endpoint contract defined but backend implementation is pending.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
        - code
        - message
        - correlationId
      properties:
        code:
          type: string
          description: Machine-readable error code.
        message:
          type: string
          description: Human-readable summary of the error.
        hint:
          type: string
          nullable: true
          description: Optional client-facing recommendation.
        correlationId:
          type: string
          format: uuid
          description: Correlates the request across systems.
        details:
          type: object
          nullable: true
          additionalProperties: true
          description: Additional contextual information for debugging.
...
    VideoPowerState:
      type: string
      enum:
        - on
        - standby
    VideoPreviewSession:
      type: object
      required:
        - sessionId
        - streamUrl
        - expiresAt
      properties:
        sessionId:
          type: string
        streamUrl:
          type: string
          format: uri
        expiresAt:
          type: string
          format: date-time
        device:
          type: object
          nullable: true
          properties:
            id:
              type: string
            name:
              type: string
            input:
              type: string
    VideoClipExport:
      type: object
      required:
        - exportId
        - recordingId
        - status
        - startOffsetSeconds
        - endOffsetSeconds
        - requestedAt
        - downloadUrl
      properties:
        exportId:
          type: string
        recordingId:
          type: string
        status:
          type: string
          enum:
            - queued
            - processing
            - complete
            - failed
        startOffsetSeconds:
          type: number
          format: float
        endOffsetSeconds:
          type: number
          format: float
        requestedAt:
          type: string
          format: date-time
        downloadUrl:
          type: string
          format: uri
    VideoRecordingSegment:
      type: object
      required:
        - id
        - deviceId
        - startedAt
        - endedAt
        - durationSeconds
        - status
      properties:
        id:
          type: string
        deviceId:
          type: string
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        durationSeconds:
          type: integer
          format: int32
        status:
          type: string
          enum:
            - available
            - processing
            - failed
    VideoDeviceState:
      type: object
      required:
        - id
        - name
        - module
        - role
        - status
        - power
        - mute
        - input
        - lastUpdated
      properties:
        id:
          type: string
        name:
          type: string
        module:
          type: string
        role:
          type: string
        status:
          type: string
        power:
          $ref: '#/components/schemas/VideoPowerState'
        mute:
          type: boolean
        input:
          type: string
        lastUpdated:
          type: string
          format: date-time
    VideoOverview:
      type: object
      required:
        - devices
        - total
        - online
        - updatedAt
      properties:
        devices:
          type: array
          items:
            $ref: '#/components/schemas/VideoDeviceState'
        total:
          type: integer
          format: int32
        online:
          type: integer
          format: int32
        updatedAt:
          type: string
          format: date-time
    VideoPreviewRequest:
      type: object
      properties:
        deviceId:
          type: string
          description: Target video node for generating the preview stream.
...
paths:
...
  /video/overview:
    get:
      tags:
        - Video
      summary: Retrieve consolidated video device status.
      description: Retrieve consolidated video device status.
      operationId: getVideoOverview
      responses:
        '200':
          description: Video overview for the primary display node.
          headers:
            x-correlation-id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoOverview'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '501':
          $ref: '#/components/responses/NotImplementedError'
        '502':
          $ref: '#/components/responses/BadGatewayError'
        '504':
          $ref: '#/components/responses/GatewayTimeoutError'
        '500':
          $ref: '#/components/responses/InternalError'
  /video/devices:
    get:
      tags:
        - Video
      summary: List managed video devices and their current state.
      description: Enumerate video endpoints, including power, mute, and input state for the operator dashboard.
      operationId: listVideoDevices
      responses:
        '200':
          description: Video device inventory.
          content:
            application/json:
              schema:
                type: object
                required:
                  - devices
                  - updatedAt
                properties:
                  devices:
                    type: array
                    items:
                      $ref: '#/components/schemas/VideoDeviceState'
                  updatedAt:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalError'
  /video/devices/{deviceId}/power:
    post:
      tags:
        - Video
      summary: Set the power state of a video endpoint.
      description: Issue power on or standby commands to a managed display or switcher.
      operationId: setVideoPower
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - power
              properties:
                power:
                  $ref: '#/components/schemas/VideoPowerState'
      responses:
        '202':
          description: Power command accepted.
          content:
            application/json:
              schema:
                type: object
                required:
                  - deviceId
                  - power
                  - lastUpdated
                properties:
                  deviceId:
                    type: string
                  power:
                    $ref: '#/components/schemas/VideoPowerState'
                  lastUpdated:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalError'
  /video/devices/{deviceId}/mute:
    post:
      tags:
        - Video
      summary: Update mute state for a video device.
      description: Toggle the audio mute flag for the specified video endpoint.
      operationId: setVideoMute
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - mute
              properties:
                mute:
                  type: boolean
      responses:
        '202':
          description: Mute update accepted.
          content:
            application/json:
              schema:
                type: object
                required:
                  - deviceId
                  - mute
                  - lastUpdated
                properties:
                  deviceId:
                    type: string
                  mute:
                    type: boolean
                  lastUpdated:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalError'
  /video/devices/{deviceId}/input:
    post:
      tags:
        - Video
      summary: Switch the active input on a video device.
      description: Change the current input selection on a display or receiver.
      operationId: setVideoInput
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - input
              properties:
                input:
                  type: string
      responses:
        '202':
          description: Input change accepted.
          content:
            application/json:
              schema:
                type: object
                required:
                  - deviceId
                  - input
                  - lastUpdated
                properties:
                  deviceId:
                    type: string
                  input:
                    type: string
                  lastUpdated:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalError'
  /video/recordings/{recordingId}/export:
    post:
      tags:
        - Video
      summary: Request an export of a recording clip for download.
      description: Submit an export intent for a recorded segment, returning tracking metadata and a download URL when ready.
      operationId: exportVideoClip
      parameters:
        - name: recordingId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - startOffsetSeconds
                - endOffsetSeconds
              properties:
                startOffsetSeconds:
                  type: number
                  format: float
                endOffsetSeconds:
                  type: number
                  format: float
      responses:
        '202':
          description: Export request accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoClipExport'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalError'
  /video/recordings:
    get:
      tags:
        - Video
      summary: Retrieve recording segments for the active video source.
      description: Retrieve recording segments for the active video source.
      operationId: getVideoRecordings
      responses:
        '200':
          description: Recording segments available for playback.
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - total
                  - generatedAt
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/VideoRecordingSegment'
                  total:
                    type: integer
                    format: int32
                  generatedAt:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '501':
          $ref: '#/components/responses/NotImplementedError'
        '502':
          $ref: '#/components/responses/BadGatewayError'
        '504':
          $ref: '#/components/responses/GatewayTimeoutError'
        '500':
          $ref: '#/components/responses/InternalError'
  /video/preview:
    post:
      tags:
        - Video
      summary: Generate a signed URL for a live video preview stream.
      description: Generate a signed URL for a live video preview stream.
      operationId: generateVideoPreview
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoPreviewRequest'
      responses:
        '200':
          description: Signed preview stream URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoPreviewSession'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '501':
          $ref: '#/components/responses/NotImplementedError'
        '502':
          $ref: '#/components/responses/BadGatewayError'
        '504':
          $ref: '#/components/responses/GatewayTimeoutError'
        '500':
          $ref: '#/components/responses/InternalError'
...